[front] refactor: consolidate slash capability helpers
